requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.11,<3.0
pydantic-settings>=2.8,<3.0
PyJWT>=2.10,<3.0
sqlalchemy>=2.0,<3.0
alembic>=1.14,<2.0
psycopg[binary]>=3.2,<4.0
redis>=5.2,<6.0
pytest>=8.3,<9.0
httpx>=0.28,<1.0
